requirements.txt
aiogram
APScheduler
python-decouple
requests
asyncio
pydantic-settings
yfinance